# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog test API Project",
        "description": "To be set Later",
        "contact": {
            "email": "slaiman.aris@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "description": "User login via email and password",
                "operationId": "App\\Http\\Controllers\\Blog\\PostController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "is_featured",
                        "in": "query",
                        "description": "is featured post {1, 0}",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "post author id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sort posts, valid options {title, created_at, is_featured}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dir",
                        "in": "query",
                        "description": "Direction of post sort ASC or DESC",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostPaginationDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Blog"
                ],
                "description": "Create new blog post",
                "operationId": "App\\Http\\Controllers\\Blog\\PostController::create",
                "requestBody": {
                    "$ref": "#/components/requestBodies/BlogPostRequest"
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "put": {
                "tags": [
                    "Blog"
                ],
                "description": "Update user blog post",
                "operationId": "App\\Http\\Controllers\\Blog\\PostController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/BlogPostRequest"
                },
                "responses": {
                    "200": {
                        "description": "Blog post updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Blog"
                ],
                "description": "Delete user post",
                "operationId": "App\\Http\\Controllers\\Blog\\PostController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blog post deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateResponseDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserDTO": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                },
                "type": "object"
            },
...
            "PostDTO": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "is_featured": {
                        "type": "boolean"
                    },
                    "comments_count": {
                        "type": "int"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserDTO"
                    }
                },
                "type": "object"
            },
            "PostPaginationDTO": {
                "properties": {
                    "total": {
                        "type": "int"
                    },
                    "current_page": {
                        "type": "int"
                    },
                    "last_page": {
                        "type": "int"
                    },
                    "per_page": {
                        "type": "int"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PostDTO"
                        }
                    }
                },
                "type": "object"
            },
            "UpdateResponseDTO": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "success": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
...
            "BlogPostRequest": {
                "description": "Blog post information to add or update",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "title": {
                                    "description": "Class BlogPostRequest",
                                    "type": "string"
                                },
                                "is_featured": {
                                    "description": "Class BlogPostRequest",
                                    "type": "boolean"
                                },
                                "content": {
                                    "description": "Class BlogPostRequest",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}